+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Added
 - Fire Baguette: Spawns a ring of fire around you and gives you the Fire Resistance effect.
 - Air Baguette: Gives you the Slow Falling effect and the Jump Boost effect.
+- Earth Baguette: Gives you the Haste effect.
 
 ## Changed
 - Water Baguette now gives Water Breathing effect.